loading and empty state GM Stat

+++ tkpdseller/src/main/java/com/tokopedia/seller/goldmerchant/statistic/view/helper/GMTransactionGraphViewHelper.java
@@ -183,7 +183,7 @@ public class GMTransactionGraphViewHelper extends BaseGMViewHelper<GMTransaction
     private void setHeaderValue(GMGraphViewWithPreviousModel data) {
         gmTitleCardView.setTitle(gmStatTransactionEntries[gmStatGraphSelection]);
 
-        gmLineChartContainer.setPercentage((double) (data.percentage * 100));
+        gmLineChartContainer.setPercentage(data.percentage * 100);
         gmLineChartContainer.setAmount(Integer.toString(data.amount));
     }
